Adding check for pid file before deletion

+++ app/Traits/DaemonTrait.php
@@ -243,7 +243,9 @@ trait DaemonTrait
         }
         // If there are processes still running, force them to quit
         exec('killall nwws.py');
-        unlink(storage_path() . '/logs/nwws.pid');
+        if (file_exists(storage_path() . '/logs/nwws.pid')) {
+            unlink(storage_path() . '/logs/nwws.pid');
+        }
         $running = FALSE;
         // Return
         if (!$running) {
